Back-end/AITrip/requirements.txt
# Django 核心框架
Django==5.2

# Django REST Framework 用于构建 API
djangorestframework==3.16.0

# Django CORS 头部处理，用于前后端分离
django_cors_headers==4.7.0

# Django 过滤器，用于 API 过滤功能
django_filter==25.1

# Django REST Framework SimpleJWT，用于 JWT 认证
djangorestframework_simplejwt==5.5.0


# MySQL 数据库驱动
mysqlclient==2.2.4
